Print the types of various data types when executed

+++ 0x00-hello_world/6-size.c
@@ -15,9 +15,9 @@ int main(void)
 	long long int size_of_long_long_int;
 	float size_of_float;
 
-	printf("Size of a char: %d byte(s)\n",(int)sizeof(size_of_char));
-	printf("Size of an int: %d byte(s)\n",(int)sizeof(size_of_an_int));
-	printf("Size of a long int: %d byte(s)\n",(int)sizeof(size_of_long_int));
-	printf("Size of a long long int: %d byte(s)\n",(int)sizeof(size_of_long_long_int));
-	printf("Size of float: %d byte(s)\n",(int)sizeof(size_of_float));
+	printf("Size of a char: %i byte(s)\n",(unsigned int)sizeof(size_of_char));
+	printf("Size of an int: %i byte(s)\n",(unsigned int)sizeof(size_of_an_int));
+	printf("Size of a long int: %ld byte(s)\n",(long int)sizeof(size_of_long_int));
+	printf("Size of a long long int: %d byte(s)\n",(unsigned int)sizeof(size_of_long_long_int));
+	printf("Size of float: %d byte(s)\n",(unsigned int)sizeof(size_of_float));
 }
